Backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.0
sqlalchemy==2.0.27
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymongo==4.5.0
motor==3.2.0
starlette==0.36.3
email-validator==2.1.0
pytest==8.3.5
pytest-asyncio==0.21.1
httpx==0.27.0
openai==1.69.0
langchain==0.3.21
langchain-openai==0.3.11 